pyproject.toml
[tool.poetry]
name = "leech"
version = "1.0.0"
description = "Turn a story on certain websites into an ebook for convenient reading"
authors = ["David Lynch <kemayo@gmail.com>"]
license = "MIT License"
include = ["ebook/*", "sites/*"]

[tool.poetry.scripts]
leech = "leech:cli"

[tool.poetry.dependencies]
python = "^3.7"
beautifulsoup4 = "^4.9.3"
click-default-group = "^1.2.2"
click = "^7.1.2"
html5lib = "^1.1"
requests = "^2.24.0"
requests-cache = "^0.5.2"
xmltodict = "*"
Pillow = "^9.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
